Add files via upload
Loan Status Prediction Analysis and Model Implementation

This notebook provides a comprehensive exploratory data analysis (EDA) and predictive modeling approach to assess loan statuses. The analysis begins with a detailed investigation of the dataset, using visualizations and statistical measures to uncover significant insights and trends among the features. Following the EDA, five distinct machine learning models are implemented and evaluated to predict loan status accurately. The models’ performances are compared using relevant metrics, and the best-performing model is identified. This project serves as a valuable resource for understanding factors influencing loan approvals and for developing robust predictive models in finance.